teams-copilot/src/requirements.txt
python-dotenv
aiohttp
teams-ai>=1.6.0,<2.0.0
colorama>=0.4.4
tavily-python>=0.3.0
requests>=2.28.0